+++ arc/utils/__init__.py
@@ -12,6 +12,7 @@ from .hooks import (
     owner_only,
     user_limiter,
 )
+from .loops import CronLoop, IntervalLoop, cron_loop, interval_loop
 from .ratelimiter import RateLimiter, RateLimiterExhaustedError
 
 __all__ = (
@@ -29,6 +30,10 @@ __all__ = (
     "LimiterHook",
     "RateLimiter",
     "RateLimiterExhaustedError",
+    "IntervalLoop",
+    "interval_loop",
+    "CronLoop",
+    "cron_loop",
 )
 
 # MIT License